Don't use redis-shutdown under systemd
The existing script calls redis-shutdown against /etc/redis.conf. It
could be fixed to use the correct server instance but the script only
supports TCP and not UNIX sockets.

There is no point in using redis-shutdown anyway as sending the TERM
signal does the same job. This is what systemd does by default.

This changes the stop from being synchronous to asynchronous but that
doesn't matter because systemd is running redis-server in the
foreground. If the shutdown operation takes too long, the server will
be killed after TimeoutStopSec either way.

There is also the added benefit that if the server has crashed, a
failed call to redis-shutdown won't prevent the server from being
started again.